use `shop` param from the request to build url
The oauth domain is not a constant with shopify - each store hais on its own subdomain.
The current setup of defining a `subdomain` in the services config array will not work when for example installing an app or sales channel from the Shopify app store. The subdomain value need to be dynamic from the `shop` GET parameter.

This PR swaps out the current `subdomain` config item for the `shop` param in the current request.
I don't think this is the ideal solution, but it works and at least gets the ball rolling to sort out the issue.

I have implemented it to have no breaking changes - if the `subdomain` is set - that will be used, if not it will fallback to the `shop` param on the request.