Fixed request and prevented undefined 'realname' exception
Where the original would use a 'json()' method on the request, this was undefined. Hence, it now takes advantage of the 'json_decode()' function. Furthermore, the response I received did not contain a 'realname' entry and thus was throwing an NPE, a check is now in place to check for this field.