Fixed NPE and JSON for Flickr provider (#119)
* Fixed request and prevented undefined 'realname' exception

Where the original would use a 'json()' method on the request, this was undefined. Hence, it now takes advantage of the 'json_decode()' function. Furthermore, the response I received did not contain a 'realname' entry and thus was throwing an NPE, a check is now in place to check for this field.

* Fixed styling to abide guidelines

* Refactored to use Laravel's `array_get()` function

https://laravel.com/docs/5.5/helpers#method-array-get

* Removed unnecessary parameter `array_get()` function