Make SAML2 binding method configurable
When implementing SAML2 configuration to work with jumpcloud.com as Identify Provider, I realised their SSO binding method is a "POST" and not a "REDIRECT". 

Problem: Redirect method is hardcoded in the code. 

Solution : This change makes it configurable in the services.php file. It is backward compatible as it takes the current value as default.