Add config option for default algorithm JWK
The current AzureB2C provider does not contain an option to provide a default algorithm when none is provided by the provider.
With this commit an additional config key is used to setup the default algorithm. It hooks into existing functionality of the JWK library which can handle a default setting.