Update the config to match the Provider calls
Provider.php calls for keys site and key from the config. The key id different from the client_id and it needs both. When you go https://stackapps.com/apps/oauth/register it issues an Id, Key, and Secret.

Called from these lines:
* https://github.com/SocialiteProviders/Providers/blob/master/src/StackExchange/Provider.php#L80
* https://github.com/SocialiteProviders/Providers/blob/master/src/StackExchange/Provider.php#L82

Without it, it returns the array key `site` not found. Does the same for the `key`.

The site needs to be one of their providers: stackoverflow works.