Some array keys may be undefined for user
As mentionned in SocialiteProviders/Providers#333, users will not necessarily have all fields filled (givenName is not guaranteed, for instance). Had this bug occur with a new Microsoft user that was made by an admin and without a name given to them.

This PR follows the same solution as the one mentionned previously. This also resolves the PR in the read-only repo SocialiteProviders/Microsoft#3.

Currently, this bug prevents correct use of this package for us, and a PHP 7.4 backport (which we are using) would be appreciated as well.