Improve lcobucci/jwt usage in Apple provider
Using the Configuration class forces to know a signer and a key, which is irrelevant
for the purpose of the provider, but it introduces some code that looks scary if used
incorrectly.

The irrelevant classes are removed, additional exception cases are now rethrowing
as InvalidStateException. Conditions have been inverted to implement early exit.